+++ sdk/api/api-base/src/main/java/io/dataease/api/template/dto/TemplateMarketDTO.java
@@ -3,10 +3,15 @@ package io.dataease.api.template.dto;
 import io.dataease.api.template.vo.MarketCategoryVO;
 import io.dataease.api.template.vo.MarketMetasVO;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.apache.commons.lang3.ArrayUtils;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 @Data
+@NoArgsConstructor
 public class TemplateMarketDTO {
     private String id;
     private String title;
@@ -18,4 +23,12 @@ public class TemplateMarketDTO {
     private Boolean showFlag = true;
     private List<MarketCategoryVO> categories;
     private MarketMetasVO metas;
+
+    public TemplateMarketDTO(String id, String title,String themeRepo,String templateUrl,String categoryName) {
+        this.id = id;
+        this.title = title;
+        this.categories =  Arrays.asList(new MarketCategoryVO(categoryName),new MarketCategoryVO("全部"));
+        this.metas = new MarketMetasVO(templateUrl);
+        this.thumbnail = themeRepo;
+    }
 }
