+++ sdk/api/api-base/src/main/java/io/dataease/api/template/response/MarketBaseResponse.java
@@ -23,4 +23,6 @@ public class MarketBaseResponse {
         this.contents = contents;
     }
 
+
+
 }
