+++ sdk/api/api-base/src/main/java/io/dataease/api/template/response/MarketMetaDataBaseResponse.java
@@ -1,0 +1,21 @@
+package io.dataease.api.template.response;
+
+import io.dataease.api.template.vo.MarketMetaDataVO;
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * Author: wangjiahao
+ * Date: 2022/7/15
+ * Description:
+ */
+@Data
+public class MarketMetaDataBaseResponse {
+
+    private List<MarketMetaDataVO> deVersion;
+
+    private List<MarketMetaDataVO> templateTypes;
+
+    private List<MarketMetaDataVO> labels;
+}
