+++ sdk/api/api-base/src/main/java/io/dataease/api/template/response/MarketTemplateV2BaseResponse.java
@@ -1,0 +1,16 @@
+package io.dataease.api.template.response;
+
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * @author : WangJiaHao
+ * @date : 2023/11/17 13:41
+ */
+@Data
+public class MarketTemplateV2BaseResponse {
+
+    private List<MarketTemplateV2ItemResult> items;
+
+}
