+++ sdk/api/api-base/src/main/java/io/dataease/api/template/response/MarketTemplateV2ItemResult.java
@@ -1,0 +1,15 @@
+package io.dataease.api.template.response;
+
+import io.dataease.api.template.vo.MarketApplicationVO;
+import lombok.Data;
+
+/**
+ * @author : WangJiaHao
+ * @date : 2023/11/17 13:41
+ */
+@Data
+public class MarketTemplateV2ItemResult {
+
+    private MarketApplicationVO application;
+
+}
