+++ sdk/api/api-base/src/main/java/io/dataease/api/template/vo/MarketApplicationSpecScreenshotBaseVO.java
@@ -1,0 +1,13 @@
+package io.dataease.api.template.vo;
+
+import lombok.Data;
+
+/**
+ * Author: wangjiahao
+ * Date: 2022/7/15
+ * Description:
+ */
+@Data
+public class MarketApplicationSpecScreenshotBaseVO {
+    private String url;
+}
