+++ sdk/api/api-base/src/main/java/io/dataease/api/template/vo/MarketApplicationSpecVO.java
@@ -1,0 +1,30 @@
+package io.dataease.api.template.vo;
+
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * Author: wangjiahao
+ * Date: 2022/7/15
+ * Description:
+ */
+@Data
+public class MarketApplicationSpecVO {
+
+    private String displayName;
+
+    private String type;
+
+    private String deVersion;
+
+    private String templateType;
+
+    private String label;
+
+    private String readmeName;
+
+    private List<MarketApplicationSpecScreenshotBaseVO> screenshots;
+
+    private List<MarketApplicationSpecLinkVO> links;
+}
