+++ sdk/api/api-base/src/main/java/io/dataease/api/template/vo/MarketMetaDataVO.java
@@ -1,0 +1,15 @@
+package io.dataease.api.template.vo;
+
+import lombok.Data;
+
+/**
+ * Author: wangjiahao
+ * Date: 2022/7/15
+ * Description:
+ */
+@Data
+public class MarketMetaDataVO {
+    private String slug;
+    private String value;
+    private String label;
+}
